moved tests from types to Id and PublicId corrected construct_id_with_keys unit test
